[ftr] Added the login feature for polly users
Now, if the user will check the checkbox for peak curation in peak
detection dialog. The user will be redirected to the login dialog,
where in the user needs to login and has to authenticate himself
and then he will be authorized for using peakMl. If the user is
already logged in through EPI, then no login would be needed again
in order to use peakMl.