[ref] Accomodate the changes of output file exported through moi
Earlier Moi exported all 88 features that were used to plot
shap figure (commonly called by- inference values). Now the change
has been made in the moi file so that it exports only 10 features
that would be represented in shap figure.
This has been done for securing the IP. As all 88 features if exposed to
user would be a problem.