[Fix] Incorrect SRM chromatograms for close pairs
When two SRM chromatgrams contain almost the exact same values for
precursor/product m/z pair (i.e., too close to be distinguished
using the existing Q1/Q3 accuracy), then the chromatogram returned
for either pair would always be the one whose first RT observation
was earlier. This was because, the filterline of the first scan
found to match the queried precursor/product pair was imposed on
all subsequent observations for the chromatogram (of that query).
This has been fixed by constructing chromatograms for all pairs
in the given range and then selecting the best among them.